build(scripts): refactor ate_abstract_account to support one account only
